added logic for hw 17, part 1 completed

+++ src/features/users/application/usecases/delete-user.command.ts
@@ -2,16 +2,16 @@ import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { UserRepository } from '../../repositories/user-repository';
 
 export class DeleteUserCommand {
-  constructor(public id: string) {}
+    constructor(public id: string) {}
 }
 
 @CommandHandler(DeleteUserCommand)
 export class DeleteUserUseCase implements ICommandHandler<DeleteUserCommand> {
-  constructor(private readonly usersRepository: UserRepository) {}
+    constructor(private readonly usersRepository: UserRepository) {}
 
-  async execute(command: DeleteUserCommand): Promise<boolean> {
-    // const user = await this.usersRepository.readUserById(command.id);
-    // if (!user) return false;
-    return this.usersRepository.deleteUser(command.id);
-  }
+    async execute(command: DeleteUserCommand): Promise<boolean> {
+        const user = await this.usersRepository.readUserById(command.id);
+        if (!user) return false;
+        return this.usersRepository.deleteUser(command.id);
+    }
 }
